Write element truth table to exodus to improve performance
The Exodus API allows for an element truth table to optionally be
sent to the Exodus library before any element data is written.  The
truth table simply tells which variables exist on which blocks.
Sending this truth table to Exodus allows for memory to be allocated
in advance, making for much more efficient writing of data,
especially if there is a large number of element blocks.

git-svn-id: https://libmesh.svn.sourceforge.net/svnroot/libmesh/trunk/libmesh@5416 434f946d-2f3d-0410-ba4c-cb9f52fb0dbf